mGridSpacingY and mGridSpacingX are now floats, so we have to type-cast to int when calculating the cell index from them. I seem to have missed Cornelius' new code when I applied my patch.
svn path=/trunk/kdepim/; revision=248770

+++ koagenda.cpp
@@ -451,10 +451,10 @@ bool KOAgenda::eventFilter_key( QObject *, QKeyEvent *ke )
         if ( !mTypeAhead ) {
           mTypeAhead = true;
           if ( mSelectionHeight > 0 ) {
-            emit newEventSignal( mSelectionCellX, mSelectionYTop / mGridSpacingY,
+            emit newEventSignal( mSelectionCellX, (int)(mSelectionYTop / mGridSpacingY),
                                  mSelectionCellX,
-                                 ( mSelectionYTop + mSelectionHeight ) /
-                                 mGridSpacingY );
+                                 (int)( ( mSelectionYTop + mSelectionHeight ) /
+                                 mGridSpacingY ) );
           } else {
             emit newEventSignal();
           }
